fix(ppo): upcast bf16 returns before numpy in histogram diagnostic
The reduced-precision buffer stores returns as bf16, but the record_histograms
path calls .numpy() on mb_returns and numpy has no bfloat16 dtype, so a real
max-autotune run crashed with "Got unsupported ScalarType BFloat16" on the first
histogram-interval update. Upcast to fp32 before the numpy conversion.

Only the record_histograms=True path hits this, which no unit test or the smoke
run exercised — add test_update_with_histograms_handles_bf16_returns to drive one
update with histograms on (reproduces the crash without the fix, CPU-only).

+++ src/boost_and_broadside/train/rl/ppo.py
@@ -1735,8 +1735,9 @@ class PPOTrainer(CheckpointMixin, LoggingMixin, OpponentMixin):
                     # Sampled from the last micro-batch of the last primary
                     # minibatch — a large-enough sample for the histograms.
                     alive_flat = hist_diag["alive_flat"]
+                    # returns are bf16-stored; upcast before numpy (no bf16 dtype there).
                     last_returns_np = (
-                        hist_diag["mb_returns"].reshape(-1, K)[alive_flat].cpu().numpy()
+                        hist_diag["mb_returns"].reshape(-1, K)[alive_flat].float().cpu().numpy()
                     )
                     last_logprob_np = hist_diag["logprob_flat"][alive_flat].cpu().numpy()
